+++ src/main/kotlin/sonique/intellij/openprojects/action/OpenProjectsAction.kt
@@ -1,4 +1,4 @@
-package sonique.intellij.recentproject.action
+package sonique.intellij.openprojects.action
 
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
@@ -6,11 +6,12 @@ import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.popup.PopupChooserBuilder
 import com.intellij.ui.components.JBList
-import sonique.intellij.recentproject.service.ProjectHolderService
+import sonique.intellij.MyBundle
+import sonique.intellij.openprojects.service.ProjectHolderService
 import javax.swing.DefaultListModel
 import javax.swing.JList
 
-class RecentProjectsAction : AnAction() {
+class OpenProjectsAction : AnAction() {
     private val projects = service<ProjectHolderService>()
 
     override fun actionPerformed(e: AnActionEvent) {
@@ -23,15 +24,14 @@ class RecentProjectsAction : AnAction() {
                 .forEach { p -> listModel.addElement(p) }
 
         val projectList: JList<Project> = JBList<Project>(listModel)
-        projectList.cellRenderer = RecentProjectsRenderer(projects.longestProjectNameLength())
+        projectList.cellRenderer = OpenProjectsRenderer(projects.longestProjectNameLength())
         projectList.selectedIndex = if (listModel.size() > 1) 1 else 0
 
         val popup = PopupChooserBuilder<Project>(projectList)
-                .setTitle("Recent Projects")
+                .setTitle(MyBundle.message("open.projects"))
                 .setItemChoosenCallback(SelectProjectRunnable(projectList))
                 .createPopup()
 
         popup.showCenteredInCurrentWindow(currentProject)
     }
 }
-