+++ src/main/kotlin/sonique/intellij/openprojects/action/OpenProjectsRenderer.kt
@@ -1,4 +1,4 @@
-package sonique.intellij.recentproject.action
+package sonique.intellij.openprojects.action
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.IconLoader
@@ -7,7 +7,7 @@ import com.intellij.ui.SimpleTextAttributes
 import java.awt.Font
 import javax.swing.JList
 
-internal class RecentProjectsRenderer(private val longestProjectName: Int) : ColoredListCellRenderer<Project>() {
+internal class OpenProjectsRenderer(private val longestProjectName: Int) : ColoredListCellRenderer<Project>() {
     override fun customizeCellRenderer(list: JList<out Project>, project: Project, index: Int, selected: Boolean, hasFocus: Boolean) {
         icon = IconLoader.findIcon("/nodes/ideaProject.png")
         font = Font("Menlo", list.font.style, list.font.size)